refactoring of functions add, getAll-add pagination, contact filtering by favorite field
